backend/db/mongo.py
import os
from pymongo import MongoClient
import certifi
from dotenv import load_dotenv

load_dotenv()

MONGODB_URI = os.getenv("MONGODB_URI")
client = MongoClient(MONGODB_URI, tls=True, tlsCAFile=certifi.where())
db = client["finance_interviewer"]
