backend/requirements.txt
fastapi
uvicorn
openai
pymongo
python-dotenv
pymupdf
pydantic
httpx
starlette
tqdm
python-multipart

